refactor: rename forms component folder in line with develop
